requirements.txt
email-validator>=2.2.0
flask>=3.1.1
flask-sqlalchemy>=3.1.1
gradio>=5.29.1
gunicorn>=23.0.0
pillow>=11.2.1
psycopg2-binary>=2.9.10
torch>=2.7.0
torchvision>=0.22.0
python-dotenv>=1.0.0